Make initialization cleaner with backend dependant variables in a backend specific file.
